[eas-cli] printCredentialsBeta -> printCredentials
