+++ packages/eas-cli/src/credentials/android/utils/__tests__/printCredentials-test.ts
@@ -4,7 +4,7 @@ import Log from '../../../../log';
 import { testAndroidAppCredentialsFragment } from '../../../__tests__/fixtures-android';
 import { createCtxMock } from '../../../__tests__/fixtures-context';
 import { getAppLookupParamsFromContext } from '../../actions/BuildCredentialsUtils';
-import { displayAndroidAppCredentials } from '../printCredentialsBeta';
+import { displayAndroidAppCredentials } from '../printCredentials';
 
 jest.mock('../../../../log');
 jest.mock('chalk', () => ({ bold: jest.fn(log => log) }));
